# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Personal Expense Tracker API - Learning Project
  description: REST API for authentication and expense tracking with Cloudinary uploads. This application is developed for learning purposes only. Do not use in production.
  x-authors:
    - Yogarajah Achchuthan
  license:
    name: ISC
  version: "1.0.2"
  x-intro:
    title: Mock Study Endpoints
    detail: |
      For quick network call practice (no auth), use these mock endpoints:
      - GET /mock/expenses → returns 5 sample expenses
      - POST /mock/expenses → echoes posted expense with a random id
servers:
  - url: http://localhost:3000/api
    description: Local
  - url: https://vertex.achchuthan.lk/api
    description: Production
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        avatar_url:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
          properties:
            user:
              $ref: "#/components/schemas/User"
            token:
              type: string
    Expense:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        amount:
          type: number
          format: float
        category:
          type: string
          enum: [food, transport, shopping, bills, entertainment, health, other]
        description:
          type: string
          nullable: true
        date:
          type: string
          format: date
        receipt_url:
          type: string
          nullable: true
        receipt_public_id:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ExpenseInput:
      type: object
      required: [amount, category, date]
      properties:
        amount:
          type: number
          format: float
          minimum: 0.01
        category:
          type: string
          enum: [food, transport, shopping, bills, entertainment, health, other]
        description:
          type: string
          maxLength: 500
        date:
          type: string
          format: date
        receipt:
          type: string
          format: binary
    ExpenseList:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            expenses:
              type: array
              items:
                $ref: "#/components/schemas/Expense"
            pagination:
              type: object
              properties:
                currentPage:
                  type: integer
                totalPages:
                  type: integer
                totalCount:
                  type: integer
                limit:
                  type: integer
...
    MockExpense:
      type: object
      properties:
        id:
          type: string
        amount:
          type: number
        category:
          type: string
        description:
          type: string
        date:
          type: string
          format: date-time
    Post:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        author_name:
          type: string
        author_email:
          type: string
          format: email
        title:
          type: string
        content:
          type: string
        comments_count:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    PostInput:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 200
        content:
          type: string
          minLength: 3
          maxLength: 5000
...
    PostList:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
          properties:
            posts:
              type: array
              items:
                $ref: "#/components/schemas/Post"
            pagination:
              type: object
              properties:
                currentPage:
                  type: integer
                totalPages:
                  type: integer
                totalCount:
                  type: integer
                limit:
                  type: integer
...
    Product:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        seller_name:
          type: string
        seller_email:
          type: string
          format: email
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
          format: float
        category:
          type: string
        image_url:
          type: string
          nullable: true
        stock:
          type: integer
        status:
          type: string
          enum: [active, inactive]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ProductInput:
      type: object
      required: [name, price, category]
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 255
        description:
          type: string
          maxLength: 2000
        price:
          type: number
          format: float
          minimum: 0.01
        category:
          type: string
          minLength: 2
          maxLength: 100
        imageUrl:
          type: string
          format: uri
        stock:
          type: integer
          minimum: 0
          maximum: 1000000
        status:
          type: string
          enum: [active, inactive]
    ProductList:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
          properties:
            products:
              type: array
              items:
                $ref: "#/components/schemas/Product"
            pagination:
              type: object
              properties:
                currentPage:
                  type: integer
                totalPages:
                  type: integer
                totalCount:
                  type: integer
                limit:
                  type: integer
...
paths:
  /mock/expenses:
    get:
      summary: Get mock expenses (study purpose)
      description: Returns a simple array of 5 mock expenses. No authentication required.
      tags: [Mock]
      responses:
        "200":
          description: Mock expenses response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      expenses:
                        type: array
                        items:
                          $ref: "#/components/schemas/MockExpense"
    post:
      summary: Post mock expense (study purpose)
      description: Accepts minimal payload and returns it with a random id. No authentication required.
      tags: [Mock]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount, category, description]
              properties:
                amount:
                  type: number
                category:
                  type: string
                description:
                  type: string
                date:
                  type: string
                  format: date-time
      responses:
        "201":
          description: Echoed mock expense
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      expense:
                        $ref: "#/components/schemas/MockExpense"
  /auth/register:
    post:
      summary: Register a new user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, name]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 6
                name:
                  type: string
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
        "409":
          description: Email already registered
  /auth/login:
    post:
      summary: Login user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /expenses:
    post:
      summary: Create expense
      tags: [Expenses]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
          multipart/form-data:
            schema:
              allOf:
                - $ref: "#/components/schemas/ExpenseInput"
      responses:
        "201":
          description: Expense created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
        "400":
          description: Validation error
        "401":
          description: Unauthorized
    get:
      summary: List expenses (paginated)
      tags: [Expenses]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Expense list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseList"
        "401":
          description: Unauthorized
...
  /posts:
    post:
      summary: Create post
      tags: [Posts]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Post created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: "#/components/schemas/Post"
        "400":
          description: Validation error
        "401":
          description: Unauthorized
    get:
      summary: List posts (paginated, optional user filter)
      tags: [Posts]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: userId
          description: Filter posts by author user id
          required: false
          schema:
            type: string
            format: uuid
        - in: query
          name: page
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Posts list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostList"
        "400":
          description: Validation error
        "401":
          description: Unauthorized
...
  /products:
    post:
      summary: Create product
      tags: [Products]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: "#/components/schemas/Product"
        "400":
          description: Validation error
        "401":
          description: Unauthorized
    get:
      summary: List products with filters
      tags: [Products]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: userId
          required: false
          schema:
            type: string
            format: uuid
        - in: query
          name: search
          required: false
          schema:
            type: string
            minLength: 1
            maxLength: 100
        - in: query
          name: category
          required: false
          schema:
            type: string
            minLength: 2
            maxLength: 100
        - in: query
          name: minPrice
          required: false
          schema:
            type: number
            minimum: 0
        - in: query
          name: maxPrice
          required: false
          schema:
            type: number
            minimum: 0
        - in: query
          name: status
          required: false
          schema:
            type: string
            enum: [active, inactive]
        - in: query
          name: page
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductList"
        "400":
          description: Validation error
        "401":
          description: Unauthorized
...
security:
  - bearerAuth: []